+++ source/infrastructure/lib/framework/bundler/runtime/langchain-py-version.ts
@@ -33,6 +33,7 @@ export class LangchainPythonVersionAssetOptions extends PythonAssetOptions {
         entry = path.resolve(entry);
         const pipOptions = packagingOptions as PipInstallArguments;
         this.dockerBuild = new LangChainPythonVersionDockerBuild(pipOptions);
+        this.localBuild = new LangChainPythonVersionLocalBuild(pipOptions);
 
         return {
             ...(assetHash && { assetHash: assetHash, assetHashType: cdk.AssetHashType.CUSTOM }),
@@ -63,9 +64,10 @@ export class LangChainPythonVersionDockerBuild extends PythonDockerBuild {
         if (process.env.SKIP_PRE_BUILD?.toLowerCase() === 'true') {
             commandList.push('python3 -m pip install poetry --upgrade');
         }
-        commandList.push(
-            `poetry run pip install --python-version ${this.evaluatedPipOptions.pythonVersion} --platform ${this.evaluatedPipOptions.platform} --implementation ${this.evaluatedPipOptions.implementation} --only-binary=${this.evaluatedPipOptions.onlyBinary} -t ${outputDir}/ dist/*.whl`
-        );
+        commandList.push('python3 -m pip install poetry-plugin-export --upgrade');
+        commandList.push(`poetry export -f requirements.txt --output ${outputDir}/requirements.txt --without-hashes`);
+        commandList.push(`poetry run pip install -r ${outputDir}/requirements.txt --python-version ${this.evaluatedPipOptions.pythonVersion} --platform ${this.evaluatedPipOptions.platform} --implementation ${this.evaluatedPipOptions.implementation} --only-binary=${this.evaluatedPipOptions.onlyBinary} -t ${outputDir}/`);
+        commandList.push(`poetry run pip install --no-deps -t ${outputDir}/ dist/*.whl`);
         return commandList;
     }
 }
@@ -81,8 +83,10 @@ export class LangChainPythonVersionLocalBuild extends PythonLocalBuild {
     protected postBuild(moduleName: string, outputDir: string): string[] {
         return [
             `cd ${moduleName}`,
-            'python3 -m pip install poetry --upgrade',
-            `poetry run pip install --python-version ${this.evaluatedPipOptions.pythonVersion} --platform ${this.evaluatedPipOptions.platform} --implementation ${this.evaluatedPipOptions.implementation} --only-binary=${this.evaluatedPipOptions.onlyBinary} -t ${outputDir}/ dist/*.whl`
+            'python3 -m pip install poetry poetry-plugin-export --upgrade',
+            `poetry export -f requirements.txt --output ${outputDir}/requirements.txt --without-hashes`,
+            `poetry run pip install -r ${outputDir}/requirements.txt --python-version ${this.evaluatedPipOptions.pythonVersion} --platform ${this.evaluatedPipOptions.platform} --implementation ${this.evaluatedPipOptions.implementation} --only-binary=${this.evaluatedPipOptions.onlyBinary} -t ${outputDir}/`,
+            `poetry run pip install --no-deps -t ${outputDir}/ dist/*.whl`
         ];
     }
 }